@@ -1,6 +1,6 @@
 {
   "name": "parse-dashboard",
-  "version": "5.0.0",
+  "version": "5.1.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/ParsePlatform/parse-dashboard"
@@ -35,15 +35,15 @@
     "LICENSE"
   ],
   "dependencies": {
-    "@babel/runtime": "7.19.0",
-    "@babel/runtime-corejs3": "7.19.1",
+    "@babel/runtime": "7.20.13",
+    "@babel/runtime-corejs3": "7.20.13",
     "bcryptjs": "2.3.0",
     "body-parser": "1.20.0",
     "commander": "9.4.0",
     "connect-flash": "0.1.1",
     "cookie-session": "2.0.0",
     "copy-to-clipboard": "3.3.2",
-    "core-js": "3.25.3",
+    "core-js": "3.28.0",
     "csurf": "1.11.0",
     "express": "4.18.1",
     "graphiql": "2.0.8",
@@ -69,17 +69,17 @@
     "react-json-view": "1.21.3",
     "react-popper-tooltip": "4.4.2",
     "react-router-dom": "6.4.1",
-    "regenerator-runtime": "0.13.9",
+    "regenerator-runtime": "0.13.11",
     "semver": "7.3.7",
     "typescript": "4.8.3"
   },
   "devDependencies": {
     "@actions/core": "1.9.1",
-    "@babel/core": "7.19.3",
+    "@babel/core": "7.20.12",
     "@babel/eslint-parser": "7.19.1",
-    "@babel/plugin-proposal-decorators": "7.19.3",
-    "@babel/plugin-transform-runtime": "7.19.1",
-    "@babel/preset-env": "7.19.3",
+    "@babel/plugin-proposal-decorators": "7.20.13",
+    "@babel/plugin-transform-runtime": "7.19.6",
+    "@babel/preset-env": "7.20.2",
     "@babel/preset-react": "7.18.6",
     "@saithodev/semantic-release-backmerge": "2.1.2",
     "@semantic-release/changelog": "5.0.1",
@@ -90,8 +90,8 @@
     "@semantic-release/release-notes-generator": "9.0.3",
     "@types/jest": "29.1.1",
     "all-node-versions": "11.0.1",
-    "babel-loader": "8.2.5",
-    "css-loader": "6.7.1",
+    "babel-loader": "9.1.2",
+    "css-loader": "6.7.3",
     "eslint": "8.24.0",
     "eslint-plugin-jest": "27.0.4",
     "eslint-plugin-react": "7.31.8",
@@ -105,13 +105,13 @@
     "react-test-renderer": "16.13.1",
     "request": "2.88.2",
     "request-promise": "4.2.5",
-    "sass": "1.55.0",
-    "sass-loader": "13.0.2",
+    "sass": "1.58.1",
+    "sass-loader": "13.2.0",
     "semantic-release": "17.4.6",
     "style-loader": "3.3.1",
     "svg-prep": "1.0.4",
-    "webpack": "5.64.4",
-    "webpack-cli": "4.9.1",
+    "webpack": "5.75.0",
+    "webpack-cli": "5.0.1",
     "yaml": "1.10.0"
   },
   "scripts": {
