@@ -54,25 +54,30 @@ function initialize(app, options) {
   });
 
   var cookieSessionSecret = options.cookieSessionSecret || require('crypto').randomBytes(64).toString('hex');
+  const cookieSessionMaxAge = options.cookieSessionMaxAge;
   app.use(require('connect-flash')());
   app.use(require('body-parser').urlencoded({ extended: true }));
   app.use(require('cookie-session')({
     key    : 'parse_dash',
     secret : cookieSessionSecret,
-    cookie : {
-      maxAge: (2 * 7 * 24 * 60 * 60 * 1000) // 2 weeks
-    }
+    maxAge : cookieSessionMaxAge
   }));
   app.use(passport.initialize());
   app.use(passport.session());
 
   app.post('/login',
     csrf(),
-    passport.authenticate('local', {
-      successRedirect: `${self.mountPath}apps`,
-      failureRedirect: `${self.mountPath}login`,
-      failureFlash : true
-    })
+      (req,res,next) => {
+        let redirect = 'apps';
+        if (req.body.redirect) {
+          redirect = req.body.redirect.charAt(0) === '/' ? req.body.redirect.substring(1) : req.body.redirect
+        }
+        return passport.authenticate('local', {
+          successRedirect: `${self.mountPath}${redirect}`,
+          failureRedirect: `${self.mountPath}login${req.body.redirect ? `?redirect=${req.body.redirect}` : ''}`,
+          failureFlash : true
+        })(req, res, next)
+    },
   );
 
   app.get('/logout', function(req, res){
